Refactor: 파일 업로드 라우터 콜백 분리
파일 업로드 라우터에서 결과를 처리하는 콜백을 controller로 분리함.
