+++ server/src/routes/file/index.js
@@ -1,0 +1,9 @@
+import express from 'express';
+import upload from '../../libs/upload';
+import FileController from './controller';
+
+const router = express.Router();
+
+router.post('/', upload.single('file'), FileController.s3UploadCallback);
+
+export default router;
